[react-relay] Split tests into modern, compat, and classic files.
