+++ types/react-relay/tsconfig.json
@@ -24,6 +24,7 @@
         "classic.d.ts",
         "compat.d.ts",
         "test/react-relay-tests.tsx",
+        "test/react-relay-compat-tests.tsx",
         "test/react-relay-classic-tests.tsx"
     ]
 }
